Create AddHighlightBoxLayer component
