+++ src/components/OpenLayersMap.tsx
@@ -2,21 +2,13 @@ import { Map, View } from 'ol';
 import { TileGrid } from 'ol/tilegrid';
 import { Tile as TileLayer } from 'ol/layer';
 import { XYZ } from 'ol/source';
-import {
-  ChangeEvent,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  useCallback,
-} from 'react';
+import { ChangeEvent, useEffect, useMemo, useRef, useState } from 'react';
 import {
   Band,
   BaselayerState,
   Box,
   SourceList,
   SubmapData,
-  NewBoxData,
 } from '../types/maps';
 import { DEFAULT_MAP_SETTINGS, SERVICE_URL } from '../configs/mapSettings';
 import 'ol/ol.css';
@@ -28,12 +20,10 @@ import VectorSource from 'ol/source/Vector';
 import './styles/highlight-controls.css';
 import './styles/area-selection.css';
 import { CropIcon } from './icons/CropIcon';
-import Draw, { createBox } from 'ol/interaction/Draw.js';
 import { HighlightBoxLayer } from './layers/HighlightBoxLayer';
-import { AddBoxDialog } from './AddBoxDialog';
-import { BoxMenu } from './BoxMenu';
 import { GraticuleLayer } from './layers/GraticuleLayer';
 import { SourcesLayer } from './layers/SourcesLayer';
+import { AddHighlightBoxLayer } from './layers/AddHighlightBoxLayer';
 
 export type MapProps = {
   bands: Band[];
@@ -68,16 +58,10 @@ export function OpenLayersMap({
 }: MapProps) {
   const mapRef = useRef<Map | null>(null);
   const drawBoxRef = useRef<VectorLayer | null>(null);
-  const drawRef = useRef<Draw | null>(null);
   const [coordinates, setCoordinates] = useState<number[] | undefined>(
     undefined
   );
-  const [newBoxData, setNewBoxData] = useState<NewBoxData | undefined>(
-    undefined
-  );
-  const [showNewBoxMenu, setShowNewBoxMenu] = useState(false);
   const [isDrawing, setIsDrawing] = useState(false);
-  const [showAddBoxDialog, setShowAddBoxDialog] = useState(false);
 
   const { activeBaselayer, cmap, cmapValues } = baselayerState;
 
@@ -186,91 +170,6 @@ export function OpenLayersMap({
     }
   }, [activeBaselayer, cmap, cmapValues, tileLayers]);
 
-  /**
-   * Handles adding/removing Draw interaction in response to isDrawing state
-   */
-  useEffect(() => {
-    const map = mapRef.current;
-    const source = drawBoxRef.current?.getSource();
-    if (map && source) {
-      if (isDrawing) {
-        source.clear();
-        setNewBoxData(undefined);
-        setShowNewBoxMenu(false);
-
-        const draw = new Draw({
-          source: source,
-          type: 'Circle',
-          geometryFunction: createBox(),
-        });
-
-        draw.on('drawend', (e) => {
-          const extent = e.feature.getGeometry()?.getExtent();
-
-          if (extent) {
-            const [
-              top_left_ra,
-              bottom_right_dec,
-              bottom_right_ra,
-              top_left_dec,
-            ] = extent;
-            const topLeftBoxPosition = map.getPixelFromCoordinate([
-              top_left_ra,
-              top_left_dec,
-            ]);
-            const bottomRightBoxPosition = map.getPixelFromCoordinate([
-              bottom_right_ra,
-              bottom_right_dec,
-            ]);
-            const boxWidth = Math.abs(
-              topLeftBoxPosition[0] - bottomRightBoxPosition[0]
-            );
-            const boxHeight = Math.abs(
-              topLeftBoxPosition[1] - bottomRightBoxPosition[1]
-            );
-            setNewBoxData({
-              top_left_ra,
-              top_left_dec,
-              bottom_right_ra,
-              bottom_right_dec,
-              top: topLeftBoxPosition[1],
-              left: topLeftBoxPosition[0],
-              width: boxWidth,
-              height: boxHeight,
-            });
-          }
-
-          setIsDrawing(false);
-        });
-
-        map.addInteraction(draw);
-        drawRef.current = draw;
-      } else {
-        if (drawRef.current) {
-          map.removeInteraction(drawRef.current);
-          drawRef.current = null;
-          setNewBoxData(undefined);
-          setShowNewBoxMenu(false);
-        }
-      }
-
-      return () => {
-        if (drawRef.current) {
-          map.removeInteraction(drawRef.current);
-          drawRef.current = null;
-        }
-      };
-    }
-  }, [mapRef.current, isDrawing]);
-
-  const handleAddBoxCleanup = useCallback(() => {
-    setNewBoxData(undefined);
-    const source = drawBoxRef.current?.getSource();
-    if (source) {
-      source.clear();
-    }
-  }, [drawBoxRef.current]);
-
   return (
     <div id="map" style={{ cursor: isDrawing ? 'crosshair' : 'auto' }}>
       <div className="ol-zoom ol-control draw-box-btn-container">
@@ -297,39 +196,15 @@ export function OpenLayersMap({
         setBoxes={setBoxes}
         submapData={submapData}
       />
-      {newBoxData && (
-        <BoxMenu
-          isNewBox={true}
-          boxData={newBoxData}
-          setShowMenu={setShowNewBoxMenu}
-          showMenu={showNewBoxMenu}
-          submapData={submapData}
-          additionalButtons={[
-            <button
-              key="add-new-box"
-              className="area-select-button highlight-box-button"
-              onClick={() => setShowAddBoxDialog(true)}
-            >
-              Add as Box
-            </button>,
-            <button
-              key="remove-region"
-              className="area-select-button highlight-box-button"
-              onClick={handleAddBoxCleanup}
-            >
-              Remove Region
-            </button>,
-          ]}
-        />
-      )}
-      <AddBoxDialog
-        showAddBoxDialog={showAddBoxDialog}
-        setShowAddBoxDialog={setShowAddBoxDialog}
-        newBoxData={newBoxData}
+      <AddHighlightBoxLayer
+        mapRef={mapRef}
+        drawBoxRef={drawBoxRef}
+        isDrawing={isDrawing}
+        setIsDrawing={setIsDrawing}
+        submapData={submapData}
         setBoxes={setBoxes}
         setActiveBoxIds={setActiveBoxIds}
         addOptimisticHighlightBox={addOptimisticHighlightBox}
-        handleAddBoxCleanup={handleAddBoxCleanup}
       />
       <LayerSelector
         bands={bands}
